pyproject.toml
[project]
name = "eden-prueba-tecnica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "dotenv>=0.9.9",
    "google-genai>=0.3.0",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "unidecode>=1.4.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
